+++ test-function-call.R
@@ -1,7 +1,7 @@
-rows <- 10
+rows <- 80
 model <- "mnl"
 efficiency_criteria <- "d-error"
-algorithm <- "rsc"
+algorithm <- "federov"
 draws <- "pseudo-random"
 R <- 100
 dudx <- "x3"
@@ -9,10 +9,10 @@ candidate_set <- NULL
 exclusions <- NULL
 control <- list(
   cores = 1,
-  max_iter = 10000,
+  max_iter = 100000,
   max_relabel = 10000,
   max_no_improve = 100000,
-  efficiency_threshold = 0.1,
+  efficiency_threshold = 0.00001,
   sample_with_replacement = FALSE
 )
 return_all <- FALSE
@@ -29,20 +29,21 @@ rm(list = ls(all = TRUE))
 #' Specifying a utility function with 3 attributes and a constant for the
 #' SQ alternative. The design has 20 rows.
 utility <- list(
-  alt1 = "b_x1_dummy[c(0, 0)] * x1[1:3] + b_x2_dummy[c(0, 0)] * x2[1:3] + b_x3_dummy[c(0, 0)] * x3[1:3]",
-  alt2 = "b_x1_dummy          * x1      + b_x2_dummy          * x2      + b_x3_dummy          * x3"
+  alt1 = "b_x1[0.1] * x1[1:6] + b_x2_dummy[c(0, 0)] * x2[1:3] + b_x3[-0.01] * x3[seq(1, 10, 2)]",
+  alt2 = "b_x1          * x1      + b_x2_dummy          * x2      + b_x3          * x3"
 )
 
 # Generate designs ----
 design <- generate_design(utility,
-                          rows = 10,
+                          rows = 50,
                           model = "mnl",
                           efficiency_criteria = "d-error",
-                          algorithm = "rsc",
+                          algorithm = "federov",
                           draws = "scrambled-sobol",
                           control = list(
-                            max_iter = 21000,
-                            max_no_improve = 5000
+                            max_iter = 100000,
+                            max_no_improve = 5000,
+                            efficiency_threshold = 0.00001
                           ))
 
 # Add a blocking variable to the design with 4 blocks.
@@ -50,3 +51,16 @@ design <- block(design, 4)
 
 
 summary(design)
+
+
+
+utility
+rows = 20
+model = "mnl"
+efficiency_criteria = "d-error"
+algorithm = "federov"
+draws = "scrambled-sobol"
+exclusions = list(
+  "alt1_x1 == 2 & alt1_x2 == 0 & alt1_x3 == 0",
+  "alt2_x2 == 1 & alt2_x3 == 1"
+)